File Database Generation
Able to generate a list of files recursively for a directory.

This system is currently in a "just works" sort of state. Seems to be some memory issues somewhere that limit directory depth and number of files that can be managed.

For now, have included a `depth` parameter for `generate_file_database` to act as a stop gap and prevent recursion from causing issues.