New Approach and Refactoring
The different parts of the program were becoming to intertwined with
each other, blurring the line between what each file is suppossed to
contain. As a result, I've decided to do a major refactoring of the
program and isolate each individual component.

I've also decided to utilize a library, such as Crypto++, to handle all
of the hashing. CRC32 is easy enough to implement, but integrating SHA
and MD5 like I originally planned is beyond my capabilities.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>