Start on Server Initialization
Working on getting the server code off the ground and implementing TLS
using OpenSSL. Theorhetically a socket is already created, bound, and
listened on, but no error handling exists yet. There is also no way of
receiving a connection from a client.

Also started an implementation of logging, which is in a similar state
to the socket.

Completely removed all code from the config code and header file because
with the changes to the server it was throwing an error during
compilation. This will be reimplemented at some point... probably.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>