SSL Socket Creation and CMake Infrastructure
Added function for creating and SSL_CTX, then subsequently spent an hour
trying to figure out how to add OpenSSL to a CMake project.

Added OpenSSL to the project with CMake. Right now this only works on
Linux. Also added the ability to call `make runserver` to run the server
executable compiled by the project.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>